new db structure and testing
