+++ routes/account.js
@@ -1,7 +1,6 @@
 const express = require('express');
 const router = express.Router();
-const sqlite3 = require('sqlite3').verbose();
-const path = require('path');
+const db = require('../db.js');
 
 router.get('/', (req, res) => {
     // AI generated code, will test later
