+++ views/payment.pug
@@ -69,7 +69,7 @@ html
 
       .cart
         .div-item
-          img(src="/public/images/star.png", alt="item-pic")
+          img(src="/images/star.png", alt="item-pic")
           .item-details
             span.title Vega
             span.price $15.00
