+++ views/signup.pug
@@ -5,7 +5,7 @@ html
     meta(name="viewport", content="width=device-width, initial-scale=1.0")
     link(href="https://fonts.googleapis.com/css2?family=Roboto&display=swap", rel="stylesheet")
     link(href="https://fonts.googleapis.com/css2?family=Bebas+Neue&family=Dancing+Script:wght@400..700&family=Libre+Baskerville:ital,wght@0,400;0,700;1,400&family=Oswald:wght@200..700&family=Playfair+Display:ital,wght@0,400..900;1,400..900&family=Roboto+Flex:opsz,wght@8..144,100..1000&display=swap", rel="stylesheet")
-    link(rel="stylesheet", href="/styles/signUp-style.css")
+    link(rel="stylesheet", href="/public/styles/signUp-style.css")
     title SignUp | Lorem
   body
     nav.navbar
